refactor: render fonts using instanced drawing
This commit refactors the font rendering system to use instanced drawing, which significantly improves performance by reducing CPU overhead and draw calls.

The key changes include:
- A new `InstancedMesh` base class and a specialized `FontMesh` have been introduced to provide a clean and reusable pattern for instanced rendering.
- The vertex shader now generates character quads on the fly using `gl_VertexID` and performs all transformations (italics, rotation) on the GPU via per-instance attributes.
- The C++ code has been refactored to produce `FontData` (the per-instance data structure) instead of individual vertices.
- The old, inefficient `FontMesh3d` system has been completely removed.
- All call-sites for the old font rendering system have been updated to use the new instanced system.
- OpenGL state management has been centralized into `GLRenderState::apply` and `Legacy::Functions::applyRenderState` for improved code clarity and maintainability.

+++ src/opengl/legacy/FontMesh.cpp
@@ -13,13 +13,13 @@ FontMesh::FontMesh(const SharedFunctions& functions, const std::shared_ptr<FontS
     : InstancedMesh(functions)
     , m_shader(sharedShader)
 {
-    m_posAttr = m_shader->getAttribLocation("aPos");
-    m_sizeAttr = m_shader->getAttribLocation("aSize");
-    m_texTopLeftAttr = m_shader->getAttribLocation("aTexTopLeft");
-    m_texBottomRightAttr = m_shader->getAttribLocation("aTexBottomRight");
-    m_colorAttr = m_shader->getAttribLocation("aColor");
-    m_italicsAttr = m_shader->getAttribLocation("aItalics");
-    m_rotationAttr = m_shader->getAttribLocation("aRotation");
+    m_posAttr = static_cast<GLint>(m_shader->getAttribLocation("aPos"));
+    m_sizeAttr = static_cast<GLint>(m_shader->getAttribLocation("aSize"));
+    m_texTopLeftAttr = static_cast<GLint>(m_shader->getAttribLocation("aTexTopLeft"));
+    m_texBottomRightAttr = static_cast<GLint>(m_shader->getAttribLocation("aTexBottomRight"));
+    m_colorAttr = static_cast<GLint>(m_shader->getAttribLocation("aColor"));
+    m_italicsAttr = static_cast<GLint>(m_shader->getAttribLocation("aItalics"));
+    m_rotationAttr = static_cast<GLint>(m_shader->getAttribLocation("aRotation"));
 }
 
 FontMesh::~FontMesh() = default;
